Move knexfile to root
