--- api/config/knexfile.ts
@@ -1,3 +1,0 @@
-import { config } from './config';
-
-module.exports = config.database;
