+++ api/knexfile.ts
@@ -1,0 +1,3 @@
+import { config } from './src/config/config';
+
+module.exports = config.database;
