+++ api/src/database/dropDatabase.ts
@@ -1,4 +1,4 @@
-import { logger } from '../../config/logger';
+import { logger } from '../config/logger';
 import { destroyClientAndExit, setupConnectionConfig, setupDbClient } from './setupClient';
 
 export const dropDatabase = async () => {
